+++ app/src/main/java/com/simcom/ecashier/model/room/PersonToGroup.kt
@@ -41,7 +41,7 @@ import androidx.navigation.ui.NavigationUI
 import androidx.room.*
 
 @Entity(tableName = "person_to_group_table")
-class PersonToGroup(val personId: Int, val groupId: Int) {
+class PersonToGroup(val personId: Long, val groupId: Long) {
     @PrimaryKey(autoGenerate = true)
     var id = 0
